requirements.txt
boto==2.48.0
boto3==1.7.4
botocore==1.10.4
docutils==0.14
jmespath==0.9.3
Pillow==5.1.0
python-dateutil==2.6.1
s3transfer==0.1.13
six==1.11.0
